requirements.txt
aiogram==3.15.0
requests==2.32.3
googlesearch-python==1.2.5
